Convert: convert by layer, and finally merge into bootstrap
The previous conversion scheme was done at one time in the first
Prepare(). There are two main disadvantages. The first is to destroy the
logic of layer-by-layer Prepare, and the second is that it is difficult
to share blobs between OCI images with same base layers.

We modify the transformation logic to do it layer by layer,
to be consistent with the Prepare behavior. The blob filename
corresponding to each layer is the 64-bit encoded value of the chainID,
the same as the snapshot key that will be commited later.

When preparing the upper layer, we obtain corresponding blob files
according to all ancestor snapshot keys, merge all blob files to
generate a final uncompressed bootstrap, that is, image/image.boot, and
save it in the last snapshot dir of the lower layer.

Signed-off-by: zhaoshang <zhaoshangsjtu@linux.alibaba.com>